+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed an issue which caused `PostProcessingSprite` to crash if application/renderer was destroyed.
+
 ## [1.1.0] - 2021-11-20
 ### Added
 - Added support to use textures included in spritesheets when using `StandardMaterial`.